horse_trainer_ai/requirements.txt
fastapi
openai
uvicorn
python-dotenv

